--- DogeClicker/app/src/main/java/com/example/dogeclicker/view/GamePage.java
@@ -1,40 +1,0 @@
-package com.example.dogeclicker.view;
-
-
-import android.view.View;
-
-import com.example.dogeclicker.R;
-
-public class GamePage {
-
-//    public void onCreate(){
-//
-//    }
-//
-//
-//
-//    public void launchActivity(){
-//
-//    }
-//
-//    public void onUpgradeCilck(View v){
-//
-//    }
-//
-//    public void onDogeCoinClick(View v){
-//
-//    }
-//
-//    public void onCursorClick(View v){
-//
-//
-//    }
-//
-//    public void onCPUClick(View v){
-//
-//    }
-//    public void onRAMClick(View v){
-//
-//    }
-
-}
