+++ DogeClicker/app/src/main/java/com/example/dogeclicker/view/PermUpgradePage.java
@@ -8,7 +8,7 @@ import com.example.dogeclicker.R;
 
 public class PermUpgradePage extends AppCompatActivity {
 
-
+    @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.perm_upgrade_page);
@@ -16,9 +16,9 @@ public class PermUpgradePage extends AppCompatActivity {
     }
 
 
-    public void launchActivity(){
-
-    }
+//    public void backButton(View v){
+//        setContentView(R.layout.activity_main);
+//    }
 
     public void onWifiClick(View v){
