+++ DogeClicker/app/src/main/java/com/example/dogeclicker/view/StartPage.java
@@ -1,0 +1,28 @@
+package com.example.dogeclicker.view;
+
+
+import android.os.Bundle;
+import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.dogeclicker.R;
+
+public class StartPage extends AppCompatActivity {
+
+   @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.start_page);
+
+    }
+
+
+
+
+
+
+
+
+
+}
